Added YAMLParsingPageObject abstract class
- Added YAMLParsingPageObject abstract class with common methods and
attributes
- Updated ModalObject and FormObject to subclass YAMLParsingPageObject
