+++ webdriver_test_tools/pageobject/modal.py
@@ -2,11 +2,11 @@ import inspect
 import os
 from selenium.common.exceptions import NoSuchElementException
 
-from webdriver_test_tools.pageobject import utils, BasePage
+from webdriver_test_tools.pageobject import utils, BasePage, YAMLParsingPageObject
 from webdriver_test_tools.webdriver import actions
 
 
-class ModalObject(BasePage):
+class ModalObject(YAMLParsingPageObject):
     """Page object prototype for modals
 
     Subclasses should set the following attributes:
@@ -29,6 +29,8 @@ class ModalObject(BasePage):
         subclasses
     """
 
+    # TODO: doc
+    YAML_ROOT_KEY = 'modal'
     # Attribute with path to YAML file (parsed on __init__)
     YAML_FILE = None
     # Locators
@@ -37,10 +39,11 @@ class ModalObject(BasePage):
     # Optional page object for the modal body content
     MODAL_BODY_CLASS = None
 
-    def __init__(self, driver):
-        super().__init__(driver)
-        if self.YAML_FILE:
-            self.parse_yaml(self.YAML_FILE)
+    # TODO: remove
+    # def __init__(self, driver):
+    #     super().__init__(driver)
+    #     if self.YAML_FILE:
+    #         self.parse_yaml(self.YAML_FILE)
 
     def parse_yaml(self, file_path):
         """Parse a YAML representation of the modal object and set attributes
@@ -51,12 +54,14 @@ class ModalObject(BasePage):
 
         :param file_path: Full path to the YAML file
         """
-        try:
-            parsed_yaml = utils.yaml.parse_yaml_file(file_path)['modal']
-        except KeyError as e:
-            raise utils.yaml.YAMLKeyError(
-                "Missing top level 'modal' key in YAML"
-            )
+        parsed_yaml = super().parse_yaml(file_path)
+        # TODO: remove
+        # try:
+        #     parsed_yaml = utils.yaml.parse_yaml_file(file_path)['modal']
+        # except KeyError as e:
+        #     raise utils.yaml.YAMLKeyError(
+        #         "Missing top level 'modal' key in YAML"
+        #     )
         # Initialize locators
         try:
             self.MODAL_LOCATOR = utils.yaml.to_locator(parsed_yaml['modal_locator'])
